update rec
updated for hyperlink in the job title + dashboard included
