+++ opentowork/pages/job_recommendation.py
@@ -27,7 +27,7 @@ def get_latest_csv_file():
     latest_csv_file = max(csv_files_paths, key=os.path.getmtime)
     return latest_csv_file
 
-def job_item(data, skills_jd, skills_resume, jd_content, resume_content):
+def job_item(data, skills_jd, skills_resume, jd_content, resume_content, key):
     """
     Create a job item using streamlit container.
 
@@ -37,6 +37,7 @@ def job_item(data, skills_jd, skills_resume, jd_content, resume_content):
         skills_resume (list): The skills present in the resume.
         jd_content (str): The job description.
         resume_content (str): The resume content.
+        key(integer): key to update job application status
 
     Returns:
         container (streamlit.container):
@@ -50,14 +51,33 @@ def job_item(data, skills_jd, skills_resume, jd_content, resume_content):
     total_skills_required = len(job_skills_set)
     container = st.container(border=True)
     col1, col2 = container.columns([5, 1])
-    col1.subheader(data['title'])
+    col1.subheader(data['title'], anchor = data['link'])
     col1.write(data['company'])
     col1.caption(data['location'])
-    col2.link_button("Apply", data['link'])
+    col2.button('applied?', on_click = status_update, args = (data,),key=key)
     col2.progress(score, text=f"{int(score*100)}%")
     col2.write(f"{skills_present_in_resume} of {total_skills_required}\
               skills are present in your resume.")
     return container
+
+def status_update(data): 
+    """
+    The main app that creates and updates job application
+    Args:
+        data (dataframe): job information
+
+    Returns:
+        dataframe: updated job application info
+    """
+    global app_status
+
+    st.toast("You Applied! Congrats")
+    new_app = [{'Company Name': data['company'], 'Position Title': data['title'], 'Location': data['location'], 'Status': 'Applied', 'Date' : datetime.now()}]
+    app_status = pd.concat([app_status, pd.DataFrame(new_app)], ignore_index=True)
+    app_status = app_status.drop_duplicates(['Company Name', 'Position Title', 'Location', 'Status'])
+    app_status.to_csv(r'C:\Users\user\Desktop\GitHub\opentowork\app_status.csv', index = None, header=True) 
+
+    return app_status
 
 def app(skills_resume, resume_content):
     """
